signup feature has been worked

+++ src/models/role.py
@@ -3,5 +3,5 @@ from src.utils.db import db
 
 
 class Role(db.Model, RoleMixin):
-    id = db.Column(db.Integer(), primary_key=True, autoincrement=True)
+    id = db.Column(db.SmallInteger(), primary_key=True, autoincrement=True)
     name = db.Column(db.String(80), unique=True)
