+++ src/models/user.py
@@ -5,7 +5,7 @@ from src.models.user_roles import UserRoles
 
 
 class User(db.Model, UserMixin):
-    id = db.Column(db.Integer(), primary_key=True, autoincrement=True)
+    id = db.Column(db.BigInteger(), primary_key=True, autoincrement=True)
     user_name = db.Column(db.String(30), unique=True, nullable=False)
     password = db.Column(db.String(255))
     email = db.Column(db.String(100), unique=True, nullable=False)
