+++ src/models/user_roles.py
@@ -2,6 +2,6 @@ from src.utils.db import db
 
 
 class UserRoles(db.Model):
-    id = db.Column(db.Integer(), primary_key=True, autoincrement=True)
-    user_id = db.Column(db.Integer(), db.ForeignKey('user.id', ondelete='CASCADE'))
+    id = db.Column(db.BigInteger(), primary_key=True, autoincrement=True)
+    user_id = db.Column(db.BigInteger(), db.ForeignKey('user.id', ondelete='CASCADE'))
     role_id = db.Column(db.Integer(), db.ForeignKey('role.id', ondelete='CASCADE'))
